+++ examples/nextjs/src/app/page.tsx
@@ -1,3 +1,4 @@
+import { CookieBannerProps } from "@/props";
 import {
   Container,
   Footer,
@@ -12,6 +13,7 @@ import {
   TextArea,
   Modal,
   Button,
+  CookieBanner,
   TypeEnum,
   List,
 } from "@govie-ds/react";
@@ -20,29 +22,30 @@ export default function Home() {
   return (
     <div className="flex flex-col h-full">
       <Header
-        logo={{ href: '/' }}
-        tools={{ search: { action: '/search-page' } }}
+        logo={{ href: "/" }}
+        tools={{ search: { action: "/search-page" } }}
         navLinks={[
           {
-            href: '#',
-            label: 'News',
+            href: "#",
+            label: "News",
           },
           {
-            href: '#',
-            label: 'Departments',
+            href: "#",
+            label: "Departments",
           },
           {
-            href: '#',
-            label: 'Services',
+            href: "#",
+            label: "Services",
           },
         ]}
         languages={[
           {
-            href: '#',
-            label: 'Gaeilge',
+            href: "#",
+            label: "Gaeilge",
           },
         ]}
       />
+      <CookieBanner {...CookieBannerProps} />
       <Container>
         <Heading>Heading</Heading>
         <PhaseBanner level="alpha">This is a pre-release version</PhaseBanner>
@@ -57,8 +60,8 @@ export default function Home() {
         <Icon icon="thumb_up" />
         <IconButton
           icon={{
-            icon: 'send',
-            ariaLabel: 'Send',
+            icon: "send",
+            ariaLabel: "Send",
           }}
         />
         <Paragraph>
@@ -75,37 +78,37 @@ export default function Home() {
         <hr />
         <RadiosGroup
           title={{
-            value: 'Where do you live?',
+            value: "Where do you live?",
             asHeading: {
-              size: 'md',
-              as: 'h2',
+              size: "md",
+              as: "h2",
             },
           }}
           items={[
             {
-              label: 'England',
-              value: 'england',
+              label: "England",
+              value: "england",
             },
             {
-              label: 'Scotland',
-              value: 'scotland',
+              label: "Scotland",
+              value: "scotland",
             },
             {
-              label: 'Ireland',
-              value: 'ireland',
+              label: "Ireland",
+              value: "ireland",
             },
           ]}
           fieldId="uniqueId"
         />
         <TextArea
           hint={{
-            text: 'Hint: This is a helpful hint.',
+            text: "Hint: This is a helpful hint.",
           }}
           id="textarea-id"
           maxChars={50}
           label={{
-            text: 'Textarea text',
-            htmlFor: 'textarea-id',
+            text: "Textarea text",
+            htmlFor: "textarea-id",
           }}
         />